ai/requirements.txt
anthropic>=0.18.0
openai>=1.1.0
python-dotenv>=0.21.0